% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions_organized.R
\name{generate.data.binary}
\alias{generate.data.binary}
\title{Title}
\usage{
generate.data.binary(Summ.Stat, n = NULL, method = 1, set.logOR.X = NULL)
}
\arguments{
\item{Summ.Stat}{a nested list where each element is a list of summary statistics for a specific data site, formatted consistently with the output from the get.summstat.binary function}

\item{n}{the number of samples to generate, the default is using the sample size from the summary statistics}

\item{method}{specify the type of data generation method, which consistent with the summary statistics input.
Options including 1, 2, 3, corresponding to the multivariate normal thresholding for A and B, the multivariate normal thresholding only for B,
and the chain of regression methods, respectively. See more details in the manuscript.}

\item{set.logOR.X}{user can manually specify the log ODDs ratio for the exposure variable X. If NULL, it use the log ODDs ratio from the summary statistics.}
}
\value{
a dataframe with the generated survival data, where the columns are B, indicator of A, X, Y, and site
}
\description{
Title
}
