% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions_organized.R
\name{generate.data.survival}
\alias{generate.data.survival}
\title{Generating data with survival outcome}
\usage{
generate.data.survival(
  Summ.Stat,
  n = NULL,
  censtype = "simple",
  trunc = 365,
  method = 1,
  set.logHR.X = NULL
)
}
\arguments{
\item{Summ.Stat}{a nested list where each element is a list of summary statistics for a specific data site, formatted consistently with the output from the get.summstat.survival function}

\item{n}{the number of samples to generate, the default is using the sample size from the summary statistics}

\item{censtype}{specify the type of censoring, with options including 'simple', 'simplebump', 'cov', and 'covbump',
see more details in the manuscript}

\item{trunc}{the truncation time for the survival data}

\item{method}{specify the type of data generation method, which consistent with the summary statistics input.
Options including 1, 2, 3, corresponding to the multivariate normal thresholding for A and B, the multivariate normal thresholding only for B,
and the chain of regression methods, respectively. See more details in the manuscript.}

\item{set.logHR.X}{user can manually specify the log hazard ratio for the exposure variable X. If NULL, it use the log hazard ratio from the summary statistics.}
}
\value{
a dataframe with the generated survival data, where the columns are B, indicator of A, X, Y, E, and site
}
\description{
Generating data with survival outcome
}
