% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions_organized.R
\name{get.summstat.binary}
\alias{get.summstat.binary}
\title{Title}
\usage{
get.summstat.binary(Y, X, B, A, method = "all")
}
\arguments{
\item{Y}{a binary vector of outcomes}

\item{X}{a binary vector of exposures}

\item{B}{a matrix of binary covariates, with each column representing a covariate}

\item{A}{a matrix of categorical covariates, with each column representing a covariate}

\item{method}{specify the type of summary statistics abstracted corresponding to the data generation method,
with options including 'all', 1, 2, 3. 'all' will return all summary statistics can be used for generate data with three methods,
while 1, 2, 3 corresponding to the multivariate normal thresholding for A and B, the multivariate normal thresholding only for B,
and the chain of regression methods, respectively. See more details in the manuscript.}
}
\value{
a list of summary statistics
}
\description{
Title
}
