% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions_organized.R
\name{get.summstat.survival}
\alias{get.summstat.survival}
\title{Getting summary statistics for data with survival outcome}
\usage{
get.summstat.survival(
  E,
  Y,
  X,
  B,
  A,
  prescription.mode = seq(30, 365, by = 30),
  my.presc.K = 1,
  tie.method = "efron",
  method = "all",
  censtype = "simple"
)
}
\arguments{
\item{E}{a numeric vector of event times}

\item{Y}{a binary vector of outcomes}

\item{X}{a binary vector of exposures}

\item{B}{a matrix of binary covariates, with each column representing a covariate}

\item{A}{a matrix of categorical covariates, with each column representing a covariate}

\item{prescription.mode}{a numeric vector of prescription modes for exposures}

\item{my.presc.K}{a numeric value of the number of prescription modes to be used}

\item{tie.method}{specifies the method for handling ties in the Cox proportional hazards model, with options including 'efron', 'breslow', and 'exact'}

\item{method}{specify the type of summary statistics abstracted corresponding to the data generation method,
with options including 'all', 1, 2, 3. 'all' will return all summary statistics can be used for generate data with three methods,
while 1, 2, 3 corresponding to the multivariate normal thresholding for A and B, the multivariate normal thresholding only for B,
and the chain of regression methods, respectively. See more details in the manuscript.}

\item{censtype}{specify the type of censoring, with options including 'simple', 'simplebump', 'cov', and 'covbump',
see more details in the manuscript}
}
\value{
a list of summary statistics
}
\description{
Getting summary statistics for data with survival outcome
}
