% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{summary.stat}
\alias{summary.stat}
\title{Title}
\usage{
\method{summary}{stat}(
  binary_outcome = FALSE,
  E = NULL,
  Y,
  X,
  B,
  A,
  prescription.mode = seq(30, trunc, by = 30),
  my.presc.K = 1,
  tie.method = "efron",
  interact = FALSE
)
}
\arguments{
\item{binary_outcome}{}

\item{E}{}

\item{Y}{}

\item{X}{}

\item{B}{}

\item{A}{}

\item{prescription.mode}{}

\item{my.presc.K}{}

\item{tie.method}{}

\item{interact}{}
}
\description{
Title
}
